+++ src/User.js
@@ -5,7 +5,7 @@ import { Component } from './framework';
 import Header from './components/Header';
 import UserInfo from './components/UserInfo';
 import Footer from './components/Footer';
-import { AUTH_HTTP } from './auth/auth-http-service.js';
+import { AUTH_HTTP } from './api/auth-http-service';
 
 export default class User extends Component {
 	constructor(props) {
