+++ src/components/LoginForm/LoginForm.js
@@ -2,7 +2,7 @@ import './login-form.scss';
 
 import { parseHTML } from '../../utils';
 import { Component } from '../../framework';
-import { AUTH_SERVICE } from '../../auth/auth-login-service';
+import { AUTH_SERVICE } from '../../api/auth-login-service';
 import Message from '../Message';
 
 export default class LoginForm extends Component {
